# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Restaurante Piscina",
        "description": "API para gestión de restaurante con módulos de cocina, meseros, órdenes, facturación y más",
        "contact": {
            "name": "Soporte Técnico",
            "email": "soporte@restaurante-piscina.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/table-assignments": {
            "get": {
                "tags": [
                    "Asignación de Mesas"
                ],
                "summary": "Obtener todas las asignaciones de mesas",
                "operationId": "b2e294573665431691159c011de8168d",
                "responses": {
                    "200": {
                        "description": "Lista de mesas con sus asignaciones"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/table-assignments/assign": {
            "post": {
                "tags": [
                    "Asignación de Mesas"
                ],
                "summary": "Asignar un mesero a una mesa",
                "operationId": "d66f9ed53ec7d4c65fc32da6c3d1c0a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "table_id",
                                    "waiter_id"
                                ],
                                "properties": {
                                    "table_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "waiter_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mesa asignada exitosamente"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/table-assignments/unassign": {
            "post": {
                "tags": [
                    "Asignación de Mesas"
                ],
                "summary": "Remover asignación de mesero de una mesa",
                "operationId": "4efb99e931feae6bf162549931d62a90",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "table_id"
                                ],
                                "properties": {
                                    "table_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Asignación removida exitosamente"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/table-assignments/waiter/{waiterId}": {
            "get": {
                "tags": [
                    "Asignación de Mesas"
                ],
                "summary": "Obtener mesas asignadas a un mesero específico",
                "operationId": "48cb36912fccf567030eb80ac6091ec9",
                "parameters": [
                    {
                        "name": "waiterId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mesas asignadas al mesero"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/table-assignments/bulk-assign": {
            "post": {
                "tags": [
                    "Asignación de Mesas"
                ],
                "summary": "Asignar múltiples mesas a un mesero",
                "operationId": "152aed55ca8cd1c29809748d7becdb57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "table_ids",
                                    "waiter_id"
                                ],
                                "properties": {
                                    "table_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    },
                                    "waiter_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mesas asignadas exitosamente"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/table-assignments/available-waiters": {
            "get": {
                "tags": [
                    "Asignación de Mesas"
                ],
                "summary": "Obtener lista de meseros disponibles",
                "operationId": "e541b42347707e37f6e914319efa6390",
                "responses": {
                    "200": {
                        "description": "Lista de meseros"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT obtenido del login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}